feat: lcp calculation and autofill

+++ src/fs.h
@@ -1,6 +1,8 @@
 #ifndef FS_H_
 #define FS_H_
 
+#include <stddef.h>
+
 typedef enum {
     REDIRECT_OUT, 
     REDIRECT_APPEND
@@ -17,6 +19,6 @@ void restore_stdout(int saved_stdout);
 int redirect_stderr(const char* path, RedirectType type);
 void restore_stderr(int saved_stderr);
 
-char** list_dir(const char* dir, int* out_count);
+char** list_dir(const char* dir, size_t* out_count);
 
 #endif
